requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
defusedxml==0.6.0
Flask==1.1.2
google-auth==1.20.1
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
jira==2.0.0
kubernetes==11.0.0
MarkupSafe==1.1.1
multidict==4.7.6
oauthlib==3.1.0
pbr==5.4.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==7.0.2
PyJWT==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
slackclient==2.7.3
slackeventsapi==2.2.0
typing-extensions==3.7.4.2
urllib3==1.25.10
websocket-client==0.47.0
Werkzeug==1.0.1
yarl==1.5.0
